publisher/requirements.txt
paho-mqtt
psycopg2-binary
pytz
